Refactored rest of TFD_Units
