+++ TFD_Units/Classes/Vehicles/Backpacks.hpp
@@ -1,0 +1,12 @@
+///BACKPACKS
+class B_AssaultPack_blk;
+class TFD_SacIntervention : B_AssaultPack_blk {
+	_generalMacro = "B_AssaultPack_blk";
+	scope = 2;
+	scopeCurator = 2;
+	scopeArsenal = 2;
+	displayName= "Sac d'intervention";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Units\Data\Backpack\SacIntervention.paa"};
+};
